improvement(interactorstyleunicam): change InteractorStyleUnicam to use manipulators
